@@ -1,14 +1,13 @@
 ﻿using AutoMapper;
 using GenericRepository.Dal.Entities;
 
-namespace GenericRepository.Api.Models
+namespace GenericRepository.Api.Models;
+
+public class VehicleProfile: Profile
 {
-    public class VehicleProfile: Profile
+    public VehicleProfile()
     {
-        public VehicleProfile()
-        {
-            CreateMap<VehicleDto, Vehicle>();
-            CreateMap<Vehicle, VehicleDto>();
-        }
+        CreateMap<VehicleDto, Vehicle>();
+        CreateMap<Vehicle, VehicleDto>();
     }
 }
