@@ -1,12 +1,14 @@
 ﻿using GenericRepository.Dal.Entities;
 
-namespace GenericRepository.Dal
-{
-    public class VehicleRepository : GenericRepository<Vehicle>, IVehicleRepository
-    {
-        protected readonly DatabaseContext _context;
+namespace GenericRepository.Dal;
 
-        public VehicleRepository(DatabaseContext context) : base(context)
-            => _context = context;
-    }
+public interface IVehicleRepository : IGenericRepository<Vehicle>
+{
+}
+public class VehicleRepository : GenericRepository<Vehicle>, IVehicleRepository
+{
+    protected readonly DatabaseContext _context;
+
+    public VehicleRepository(DatabaseContext context) : base(context)
+        => _context = context;
 }
